+++ lib/image_library/services/image_save_handler.dart
@@ -1,4 +1,6 @@
+import 'dart:io';
 import 'dart:typed_data';
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:image/image.dart' as img;
 import 'package:magicepaperapp/constants/color_constants.dart';
@@ -7,6 +9,7 @@ import 'package:magicepaperapp/image_library/widgets/dialogs/storage_permisson_d
 import 'package:magicepaperapp/image_library/provider/image_library_provider.dart';
 import 'package:magicepaperapp/image_library/services/image_operations_service.dart';
 import 'package:magicepaperapp/image_library/widgets/dialogs/image_save_dialog.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 class ImageSaveHandler {
   final BuildContext context;
@@ -24,7 +27,7 @@ class ImageSaveHandler {
   Future<bool> checkPermissionBeforeAction() async {
     if (_hasStoragePermission) return true;
 
-    final granted = await StoragePermissionDialog.checkAndRequestPermission(
+    final granted = await checkAndRequestPermission(
       context,
       colorAccent: colorAccent,
       colorBlack: colorBlack,
@@ -134,6 +137,93 @@ class ImageSaveHandler {
       flipHorizontal,
       flipVertical,
       modelId,
+    );
+  }
+
+  static Future<bool> checkAndRequestPermission(
+    BuildContext context, {
+    Color colorAccent = Colors.blue,
+    Color colorBlack = Colors.black,
+  }) async {
+    if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+      return true;
+    }
+    Permission permission;
+
+    if (Platform.isAndroid) {
+      final androidInfo = await DeviceInfoPlugin().androidInfo;
+      if (androidInfo.version.sdkInt >= 33) {
+        permission = Permission.photos;
+      } else {
+        permission = Permission.storage;
+      }
+    } else if (Platform.isIOS) {
+      permission = Permission.photos;
+    } else {
+      return false;
+    }
+
+    var status = await permission.status;
+    if (status.isGranted) return true;
+
+    if (status.isDenied || status.isPermanentlyDenied) {
+      if (!context.mounted) return false;
+
+      bool? userAgreed = await showDialog<bool>(
+        context: context,
+        barrierDismissible: false,
+        builder: (BuildContext dialogContext) {
+          return StoragePermissionDialog(
+            colorAccent: colorAccent,
+            colorBlack: colorBlack,
+            onGrantPermission: () {
+              Navigator.of(dialogContext).pop(true);
+            },
+            onCancel: () {
+              Navigator.of(dialogContext).pop(false);
+            },
+          );
+        },
+      );
+
+      if (userAgreed == true) {
+        status = await permission.request();
+
+        if (status.isGranted) {
+          return true;
+        }
+
+        if (status.isPermanentlyDenied && context.mounted) {
+          _showSettingsRedirectDialog(context);
+        }
+      }
+    }
+
+    return false;
+  }
+
+  static void _showSettingsRedirectDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        title: const Text('Permission Required'),
+        content: const Text(
+          'Storage permission is permanently denied. Please enable it in the app settings to continue.',
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(context).pop(),
+            child: const Text('Cancel'),
+          ),
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+              openAppSettings();
+            },
+            child: const Text('Open Settings'),
+          ),
+        ],
+      ),
     );
   }
 }
