+++ lib/image_library/widgets/dialogs/storage_permisson_dialog.dart
@@ -123,88 +123,4 @@ class StoragePermissionDialog extends StatelessWidget {
       ),
     );
   }
-
-  static Future<bool> checkAndRequestPermission(
-    BuildContext context, {
-    Color colorAccent = Colors.blue,
-    Color colorBlack = Colors.black,
-  }) async {
-    Permission permission;
-
-    if (Platform.isAndroid) {
-      final androidInfo = await DeviceInfoPlugin().androidInfo;
-      if (androidInfo.version.sdkInt >= 33) {
-        permission = Permission.photos;
-      } else {
-        permission = Permission.storage;
-      }
-    } else if (Platform.isIOS) {
-      permission = Permission.photos;
-    } else {
-      return false;
-    }
-
-    var status = await permission.status;
-    if (status.isGranted) return true;
-
-    if (status.isDenied || status.isPermanentlyDenied) {
-      if (!context.mounted) return false;
-
-      bool? userAgreed = await showDialog<bool>(
-        context: context,
-        barrierDismissible: false,
-        builder: (BuildContext dialogContext) {
-          return StoragePermissionDialog(
-            colorAccent: colorAccent,
-            colorBlack: colorBlack,
-            onGrantPermission: () {
-              Navigator.of(dialogContext).pop(true);
-            },
-            onCancel: () {
-              Navigator.of(dialogContext).pop(false);
-            },
-          );
-        },
-      );
-
-      if (userAgreed == true) {
-        status = await permission.request();
-
-        if (status.isGranted) {
-          return true;
-        }
-
-        if (status.isPermanentlyDenied && context.mounted) {
-          _showSettingsRedirectDialog(context);
-        }
-      }
-    }
-
-    return false;
-  }
-
-  static void _showSettingsRedirectDialog(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('Permission Required'),
-        content: const Text(
-          'Storage permission is permanently denied. Please enable it in the app settings to continue.',
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.of(context).pop(),
-            child: const Text('Cancel'),
-          ),
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop();
-              openAppSettings();
-            },
-            child: const Text('Open Settings'),
-          ),
-        ],
-      ),
-    );
-  }
 }
